acscope: use filter args instead of regexp
The regexp used to parse the cscope output is common to every search,
the only difference is that sometime we want to filter the result.

We can remove the regexp args and introduce a new arg: filter.
filter is a function called in `acscope-find--parse-line', if the
filter function return t, we return the result.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>